requirements.txt

requests
beautifulsoup4
pytesseract
reportlab
arabic_reshaper
Pillow
python-bidi
transformers
torch
accelerate>=0.27.2  # Required by transformers for device_map
einops>=0.6.1       # Required by some vision-language models

langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain_core
langgraph
pydantic
tavily-python
python-dotenv
huggingface-hub
sentence-transformers
chromadb
langdetect
pypdf
ipython
matplotlib
langchain-groq
langchain-chroma
fastapi
uvicorn
requests
langchain-tavily